fix per yair comments
