+++ integrations/gitlab/CHANGELOG.md
@@ -14,6 +14,10 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Added group & subgroup webhook support (PORT-6229)
 
+### Bug Fixes
+
+- Fixed a bug when checking whether a group should be synced or not (#1)
+
 
 0.1.48 (2024-01-23)
 ===================
