+++ integrations/gitlab/gitlab_integration/events/hooks/jobs.py
@@ -2,12 +2,12 @@ from typing import Any
 
 from gitlab.v4.objects import Project
 
-from gitlab_integration.events.hooks.base import HookHandler
+from gitlab_integration.events.hooks.project import ProjectHandler
 from gitlab_integration.utils import ObjectKind
 from port_ocean.context.ocean import ocean
 
 
-class Job(HookHandler):
+class Job(ProjectHandler):
     events = ["Job Hook"]
     system_events = ["job"]
